Add styles to <p>, and IDs "two" & "four"
Because the assignment requested for all odd numbers to have the same
styles, I applied the same style to all <p>, since they were used for
the odd numbers.

For the even numbers, the assignment request different styles for them.
So I made the IDs "two" and "four", and applied styles to these IDs.
I also found out that IDs cannot be numbers themselves, i.e., "2" & "4".